workspace/skills/astro_guide/requirements.txt
# AstroGuide Hardware Skill v2.0 - OpenClaw 技能依赖
# 智慧天文指星仪控制技能 + 课件自动生成功能

# ============================================
# 硬件控制模块
# ============================================
# 串口通信 - 硬件设备控制核心依赖
pyserial>=3.5

# 文字转语音 - 两种引擎可选
edge-tts>=6.1.9      # 微软 Edge 在线 TTS（推荐，音质好）
pyttsx3>=2.90        # 离线 TTS（备选，无需网络）

# ============================================
# 生产力模块 - 课件自动生成
# ============================================
# PPT 生成库 - 用于动态生成 PowerPoint 幻灯片
python-pptx>=0.6.21

# ============================================
# 自动化流水线模块
# ============================================
# YAML 解析 - 流水线文件格式
PyYAML>=6.0

# ============================================
# 天文计算模块（可选，用于更精确的坐标计算）
# ============================================
# Skyfield - 高精度天文计算
# skyfield>=1.45

# Astroquery - 在线天文数据查询
# astroquery>=0.4.6

# ============================================
# 其他依赖
# ============================================
# 异步支持
# asyncio (Python 3.7+ 内置)

# 音频播放（可选，用于跨平台兼容）
# pydub>=0.25.1

# Windows 终端颜色支持（可选）
# colorama>=0.4.6
